app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dropbox>=12.0.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
